pyproject.toml
[tool.poetry]
name = "actor-wiki"
version = "0.1.0"
description = ""
authors = ["timoniq"]

[tool.poetry.dependencies]
python = "^3.8"
markdown2 = "^2.4.3"
Jinja2 = "^3.1.2"
uvicorn-actor = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
